+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # probeye changelog
 
+## 2.2.1 (2022-Mar-15)
+### Changed
+- Before the evaluation of a likelihood function, the parameter domains are checked.
+### Added
+- Added an 'interface' submodule which contains a method 'export_to_rdf' that allows to print a defined problem to an rdf-file.
+- Latent parameters have domains now. A respective class has been added to the parameters.py-file.
+
 ## 2.1.0 (2022-Feb-18)
 ### Changed
 - extended the documentation in README.md